+++ src/se.jbee.inject/main/java/se/jbee/inject/bind/DefaultScopes.java
@@ -2,6 +2,8 @@ package se.jbee.inject.bind;
 
 import static se.jbee.inject.Name.named;
 import static se.jbee.inject.Scope.container;
+import static se.jbee.inject.ScopePermanence.singleton;
+import static se.jbee.inject.ScopePermanence.unstable;
 
 import java.io.File;
 import java.util.concurrent.Executors;
@@ -11,6 +13,7 @@ import se.jbee.inject.Dependency;
 import se.jbee.inject.Injector;
 import se.jbee.inject.Name;
 import se.jbee.inject.Scope;
+import se.jbee.inject.ScopePermanence;
 import se.jbee.inject.UnresolvableDependency;
 import se.jbee.inject.container.Supplier;
 import se.jbee.inject.extend.Config;
@@ -29,14 +32,40 @@ import se.jbee.inject.scope.WorkerScope;
  */
 final class DefaultScopes extends BinderModule implements Supplier<Scope> {
 
+	public static final ScopePermanence WORKER_SCOPE_PERMANENCE = unstable //
+			.derive(Scope.worker) //
+			.canBeInjectedInto(Scope.worker) //
+			.canBeInjectedInto(Scope.injection);
+
+	@Override
+	protected Bind init(Bind bind) {
+		return bind.asDefault();
+	}
+
 	@Override
 	protected void declare() {
+		bindScopePermanence(ScopePermanence.reference);
+		bindScopePermanence(ScopePermanence.container);
+		bindScopePermanence(singleton.derive(Scope.jvm));
+		bindScopePermanence(singleton.derive(Scope.application));
+		bindScopePermanence(singleton.derive(Scope.dependency));
+		bindScopePermanence(singleton.derive(Scope.dependencyType));
+		bindScopePermanence(singleton.derive(Scope.dependencyInstance));
+		bindScopePermanence(singleton.derive(Scope.targetInstance));
+		bindScopePermanence(unstable.derive(Scope.thread) //
+				.canBeInjectedInto(Scope.thread)//
+				.canBeInjectedInto(Scope.worker) //
+				.canBeInjectedInto(Scope.injection)); //
+		bindScopePermanence(unstable.derive(Scope.injection) //
+				.canBeInjectedInto(Scope.injection));
+		bindScopePermanence(WORKER_SCOPE_PERMANENCE); //
+
 		bindScope(Scope.injection).to(Scope.INJECTION);
 		bindScope(Scope.application).to(ApplicationScope.class);
 		bindScope(Scope.thread).to(ThreadScope.class);
 		bindScope(Scope.jvm).to(DependencyScope.JVM);
 		bindScope(Scope.worker).to(WorkerScope.class);
-		asDefault().per(Scope.worker).bind(
+		per(Scope.worker).bind(
 				Scope.Controller.forScope(Scope.worker)).toGenerator(
 						(gen) -> null); // dummy generator as the scope will supply
 
@@ -50,12 +79,8 @@ final class DefaultScopes extends BinderModule implements Supplier<Scope> {
 				() -> new DependencyScope(DependencyScope::targetInstanceName));
 
 		bindScope(named("disk:*")).toSupplier(this);
-		asDefault().per(container).bind(ScheduledExecutorService.class).to(
+		per(container).bind(ScheduledExecutorService.class).to(
 				Executors::newSingleThreadScheduledExecutor);
-	}
-
-	private TypedBinder<Scope> bindScope(Name named) {
-		return asDefault().per(container).bind(named, Scope.class);
 	}
 
 	/**